// packages/spotify-effect/src/model/SpotifyOptions.ts
export type SearchOptions = {
  market?: string;
  limit?: number;
  offset?: number;
  include_external?: "audio";
};